+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -20,6 +20,7 @@ Added
 - Added `IGameSessionEventHandler.OnGameSessionReset` event fired when the gamesession is reset by the host.
 - Added `IGameSessionEventHandler.ShouldCompleteGame` event fired to decide if the gamesession should evaluate the results posted with PostResult and complete the game.
 - Added property `CreatedOn` on `IGamesessionService` to get the UTC date the gamesession was created.
+- Added Session to context of `IGameSessionEventHandler.OnGameSessionCompleted`
 
 Fixed
 *****